Implementacion controllers de la fachada y modificaciones generales

+++ src/main/java/turnopro/utils/IdProgresivoGenerator.java
@@ -3,11 +3,11 @@ package turnopro.utils;
 import turnopro.controllers.FachadaControllers;
 
 public class IdProgresivoGenerator {
-    public static Integer generarSiguiente(FachadaControllers fachada){
-        Integer maxId = fachada.obtenerMaximoIdentificadorProgresivo();
+    public static Long generarSiguiente(FachadaControllers fachada){
+        Long maxId = fachada.obtenerMaximoIdentificadorProgresivo();
 
         if (maxId == null){
-            return 1;
+            return 1L;
         }
         return maxId + 1;
     }
